Moved details section to its own component
